+++ changelog.md
@@ -3,7 +3,7 @@
 * [ARVE Pro changelog](https://nextgenthemes.com/plugins/arve-pro/#changelog)
 * [ARVE Random Videos changelog](https://nextgenthemes.com/plugins/arve-random-video/#changelog)
 
-### 2025-03-09 10.6.3 ###
+### 2025-03-09 10.6.4 ###
 
 * Fix: Link to settings page in API key notice.
 * Improved: Link to tutorial video showing you how to get an YouTube Data API key.